feat - Support glob input
Note that I extracted the logic to a worker. Unfortunately proselint
isn't worker ready due to https://github.com/amperser/proselint/issues/313 .
