+++ package.json
@@ -24,8 +24,10 @@
     "test": "npm run lint"
   },
   "dependencies": {
+    "async": "^2.1.4",
     "babel-code-frame": "^6.22.0",
     "chalk": "^1.1.3",
+    "glob": "^7.1.1",
     "remark": "^6.2.0",
     "split-lines": "^1.1.0",
     "temp-write": "^3.0.0",
